Jenkins helper (#105)
* Change dependency from cognite-sdk to cognite-sdk-core.

We are not using pandas and numpy, so by using the core sdk
we can get rid of some dependencies and build faster.

Also snuck in tox testing on 3.8, and pinned black to specific
version.

Signed-off-by: Even Wiik Thomassen <even.thomassen@cognite.com>
* Update Jenkinsfile to remove some dev-dependencies.

Manage to get rid of twine and yolk from dev dep.
Added isort back.
Moved tox to pyproject.toml file.

Signed-off-by: Even Wiik Thomassen <even.thomassen@cognite.com>
* Remove reference to snakepit.

Signed-off-by: Even Wiik Thomassen <even.thomassen@cognite.com>
* Merge branch 'master' into jenkins-helper

Signed-off-by: Even Wiik Thomassen <even.thomassen@cognite.com>
* Run tox through poetry.

Signed-off-by: Even Wiik Thomassen <even.thomassen@cognite.com>
* Fix imports order in test of replicator.

Signed-off-by: Even Wiik Thomassen <even.thomassen@cognite.com>
* Remove poetry command from docs building.

Signed-off-by: Even Wiik Thomassen <even.thomassen@cognite.com>